[#227] Refactor and move a11y code to task file.
